xf86-video-intel: enable xvmc feature by default and fix dependency
XvMC extends the X Video extension (Xv) and enables hardware rendered
motion compensation support. In a test build enabling this feature
increased the size of the xf86-video-intel package from 1386841 to
1847154 bytes.

When we enable the xvmc feature in xf86-video-intel we see the
following QA issue reported:
QA Issue: xf86-video-intel rdepends on xcb-util, but it isn't
a build dependency?

We fix this by ensuring the build dependencies are set correctly when
the xvmc feature is enabled.

Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>