alsa-lib: Lift the limitation to use gcc only cross compiler
alsa-lib configure tries to see that if CC variabled doesnt contain
'gcc' then it tries to guess and redefine CC to be something gcc, which
doesnt work with OE since we define CC to have sysroot and options
attached to it. So all tests fail when configure constructs the CC for
cross compiler like this. This portion is not needed at all

(From OE-Core rev: 33a7eb64cded61601c6ad6883e4d64ebe9ff547c)

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>