os-release: Exclude DATETIME from sstate signature
I'm torn over this, I can see someone wanting this to vary in each
build. What pushed me to believe its incorrect in its current form
is that it changes for every MACHINE build, making a complete mess
if you try and generate package feeds using it.

The alternative would be to give up on allarch in this case and make
it MACHINE specific which at least would then be more consistent when
the package makes it to a package feed. If someone wants to do that,
the can propose another patch but this change at least make package
feed usage 'sane' again and avoids the perpetual rebuilds.

Incidentally, its worth noting that changes in source metadata revision
used for the build and included in this recipe would still trigger
rebuilds which is likely the common use case people actually care about.

(From OE-Core rev: 46de6f6eb9c4387298fafb48fb9c36abc3ff48a4)

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>