linux-yocto/3.14/4.1: fix ARM boot with gcc 5.x
Booting qemuarm when the kernel was built with gcc 5.x would result in
a boot hang (or at least no visible output).

Updating the SRCREVs for the following change:

    Author: Jianchuan Wang
    Email: jianchuan.wang@windriver.com
    Subject: Omit to optimize vsprintf.c/kasprintf.c
    Date: Wed, 26 Aug 2015 13:54:57 +0800

    Add "-O0" for vsprintf.c/kasprintf.c

    Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>

While this may not be the final fix (it still has to go usptream and
be better explained), it gets us booting, so is good enough for the
time being.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>