bind: remove dependency on bind in bind-dev
bind doesn't ship shared libraries only static libraries, so the default
dependency on PN from PN-dev is pointless and means that an image with bind-dev
installed (via dhcp-dev's automatic dependency) ends up with named installed and
started on boot which is rarely intended.

If and when we ship bind's shared libraries we should ensure that the libraries
go into a separate package.

Also remove an old comment about --enable-exportlib which isn't supported by
configure anymore.

[ YOCTO #8216 ]

Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>