tcmode-default: Set gcc 5.2 as the default
gcc 5.X is now working in all the places we test it in, its been in
testing for quite some time. Time to make it the default (we have some
room in M4 for any other bugfixes). Its easy to switch back to 4.9, we
should really remove 4.8 at this point (to meta-oe?).

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>