u-boot.inc: Add host compiler flags and openssl-native dependency
U-Boot will compile its own tools during the build, with specific
configurations (aka when CONFIG_FIT_SIGNATURE is enabled) the host tools
require openssl. This patch adds 'openssl' as a PACKAGECONFIG that
enables the use of openssl-native as a dependency and also adds the
HOSTCC flags that U-Boot uses when compiling the host tools.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>