populate_sdk_base: Fix aarch64 OLDEST_KERNEL sdk issues
aarch64 sets OLDEST_KERNEL to 3.14. This stops the aarch64 SDK installing on
anything with an older kernel which is clearly incorrect.

I attempted to extract the correct non-overridden version from the data store
but it proved problematic and I was running into data store issues. Those
are a separate problem but there isn't time to fix this right now.

Instead just code the SDK kernel version separately to work around this for
now (and fix the autobuilder tests and SDK usage).

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>