gcc-shared-source: Set empty SRC_URI
gcc-source is the only gcc recipe meant to handle the fetch/unpack/patch
tasks, the other gcc recipes then depend on this.

This approach has been creating some confusion for tools like the archiver.
The simplest way to signal to these processes that there is no source
is to empty SRC_URI at the same time we disable the other tasks.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>