udev: add PROVIDES = "libgudev"
With the last update of systemd libgudev was splitted out of systemd. To make
packages depending on libgudev happy, a recipe building libgudev was created in
meta-oe and the dependencies were modified from udev to libgudev.

This works fine for distros using systemd as init system, but distros not using
build udev which provides libgudev.

Signed-off-by: Andreas Müller <schnitzeltony@googlemail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>